Add files via upload
I'm sorry, but I couldn't read anything in the tutorial video series "Creating a basic HTML website". The only reason I could write the required code for "part 1" titled "Index.html" is because I have the basic HTML template memorized. Is there possibly a higher resolution copy of the video series, or a text document with the information? 
Thank you for your help.